Refactored to support multiple search backends
