+++ src/main/java/net/pms/network/mediaserver/handlers/LucenseSearchRequestHandler.java
@@ -16,38 +16,14 @@
  */
 package net.pms.network.mediaserver.handlers;
 
-import java.sql.Connection;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.sql.Statement;
-import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import org.apache.commons.configuration.ConfigurationException;
-import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import net.pms.configuration.UmsConfiguration;
-import net.pms.database.MediaDatabase;
-import net.pms.database.MediaTableAudioMetadata;
-import net.pms.dlna.DidlHelper;
-import net.pms.formats.Format;
-import net.pms.media.audio.metadata.AlbumMetadata;
-import net.pms.network.mediaserver.HTTPXMLHelper;
 import net.pms.network.mediaserver.handlers.message.SearchRequest;
-import net.pms.renderers.Renderer;
-import net.pms.store.DbIdLibrary;
 import net.pms.store.DbIdMediaType;
-import net.pms.store.DbIdResourceLocator;
-import net.pms.store.DbIdTypeAndIdent;
-import net.pms.store.MediaStoreIds;
-import net.pms.store.StoreResource;
-import net.pms.store.container.MusicAlbumFolder;
-import net.pms.store.container.MusicBrainzPersonFolder;
 
 /**
  * <pre>
@@ -60,118 +36,21 @@ import net.pms.store.container.MusicBrainzPersonFolder;
  * by the document <b>ContentDirectory:1 Service Template Version 1.01 Section 2.5.5.1</b>.
  * </pre>
  */
-public class SearchRequestHandler {
+public class LucenseSearchRequestHandler extends BaseSearchRequestHandler {
 
-	private static final Logger LOGGER = LoggerFactory.getLogger(SearchRequestHandler.class);
-	private static final String CRLF = "\r\n";
-	private static final Pattern CLASS_PATTERN = Pattern.compile("upnp:class\\s(\\bderivedfrom\\b|=)\\s+\"(?<val>.*?)\"",
-		Pattern.CASE_INSENSITIVE);
-	private static final Pattern ARTIST_ROLE = Pattern.compile("upnp:class.*role\\s*=\\s*\"(?<val>.*?)\".*", Pattern.CASE_INSENSITIVE);
-
-	private static final Pattern TOKENIZER_PATTERN = Pattern.compile(
-		"(?<property>((dc)|(upnp)):[A-Za-z@\\[\\]\"=]+)\\s+" +
-		"(?<op>[A-Za-z=!<>]+)\\s+\"(?<val>([^\"]|\"\")*)\"", Pattern.CASE_INSENSITIVE);
-
+	private static final Logger LOGGER = LoggerFactory.getLogger(LucenseSearchRequestHandler.class.getName());
 	private static final Pattern LUCENE_PATTERN = Pattern.compile("([-+&|!(){}\\[\\]^\"~*?:/\\\\])");
 
-	public record SearchToken(String attr, String op, String val) { }
 
-	private SearchRequest requestMessage = null;
-	private DbIdMediaType requestType = null;
-	private List<SearchToken> tokens = null;
-
-	private static UmsConfiguration umsConfiguration;
-
-	static {
-		try {
-			umsConfiguration = new UmsConfiguration();
-		} catch (ConfigurationException | InterruptedException e) {
-			LOGGER.error("Error while initializing SearchRequestHandler : ", e);
-			throw new RuntimeException(e);
-		}
+	public LucenseSearchRequestHandler(SearchRequest requestMessage) {
+		super(requestMessage);
 	}
 
-	public SearchRequestHandler(SearchRequest requestMessage) {
-		this.requestMessage = requestMessage;
-		requestType = getRequestType(requestMessage.getSearchCriteria());
-		tokens = getSearchTokens(requestMessage.getSearchCriteria());
-	}
-
-	private DbIdMediaType getRequestType(String searchCriteria) {
-		Matcher matcher = CLASS_PATTERN.matcher(searchCriteria);
-		if (matcher.find()) {
-			String propertyValue = matcher.group("val");
-			LOGGER.trace("upnp:class is {}", propertyValue);
-			if (propertyValue != null) {
-				propertyValue = propertyValue.toLowerCase();
-				// More specific types must be checked first
-				if (propertyValue.startsWith("object.item.audioitem")) {
-					return DbIdMediaType.TYPE_AUDIO;
-				} else if (propertyValue.startsWith("object.item.videoitem")) {
-					return DbIdMediaType.TYPE_VIDEO;
-				} else if (propertyValue.startsWith("object.item.imageitem")) {
-					return DbIdMediaType.TYPE_IMAGE;
-				} else if (propertyValue.startsWith("object.container.person")) {
-					return resolveRolePerson(searchCriteria);
-				} else if (propertyValue.startsWith("object.container.album")) {
-					return DbIdMediaType.TYPE_ALBUM;
-				} else if (propertyValue.startsWith("object.container.playlistcontainer")) {
-					return DbIdMediaType.TYPE_PLAYLIST;
-				} else if (propertyValue.startsWith("object.container")) {
-					return DbIdMediaType.TYPE_FOLDER;
-				}
-			}
-		}
-		throw new RuntimeException("Unknown type : " + (searchCriteria != null ? searchCriteria : "NULL"));
-	}
-
-	private DbIdMediaType resolveRolePerson(String searchCriteria) {
-		Matcher matcher = ARTIST_ROLE.matcher(searchCriteria);
-		if (matcher.find()) {
-			String roleValue = matcher.group("val");
-			if ("composer".equalsIgnoreCase(roleValue)) {
-				LOGGER.debug("looking up artist composer");
-				return DbIdMediaType.TYPE_PERSON_COMPOSER;
-			} else if ("conductor".equalsIgnoreCase(roleValue)) {
-				LOGGER.debug("looking up artist conductor");
-				return DbIdMediaType.TYPE_PERSON_CONDUCTOR;
-			} else if ("AlbumArtist".equalsIgnoreCase(roleValue)) {
-				LOGGER.debug("looking up artist AlbumArtist");
-				return DbIdMediaType.TYPE_PERSON_ALBUMARTIST;
-			}
-			LOGGER.warn("unknown artist role {}. Fallback to artist search ... ", roleValue);
-			return DbIdMediaType.TYPE_PERSON;
-		} else {
-			LOGGER.trace("artist without role. Regular artist search.");
-			return DbIdMediaType.TYPE_PERSON;
-		}
-	}
-
-	/**
-	 * Not used at this time. UmsContentDirectoryService creates the response object.
-	 * @param renderer
-	 * @return
-	 */
-	protected StringBuilder createSearchResponse(Renderer renderer) {
-		int numberReturned = 0;
-		StringBuilder dlnaItems = new StringBuilder();
-
-		int totalMatches = getLibraryResourceCountFromSQL(convertToCountSql(requestMessage));
-
-		for (StoreResource resource : getLibraryResourceFromSQL(renderer)) {
-			numberReturned++;
-			dlnaItems.append(DidlHelper.getDidlString(resource));
-		}
-
-		// Build response message
-		StringBuilder response = buildEnvelope(numberReturned, totalMatches, MediaStoreIds.getSystemUpdateId().getValue(), dlnaItems);
-		return createResponse(response.toString());
-	}
 
 	private String getFormattedLuceneString(String column, String sql, boolean ignoreCountLimit) {
-		String title = getLuceneTitleMatch(tokens);
-		int count = requestMessage.getRequestedCount() != null ? requestMessage.getRequestedCount() : 0;
-		int startIndex = requestMessage.getStartingIndex() != null ? requestMessage.getStartingIndex() : 0;
+		String title = getLuceneTitleMatch(getTokens());
+		int count = getRequestMessage().getRequestedCount() != null ? getRequestMessage().getRequestedCount() : 0;
+		int startIndex = getRequestMessage().getStartingIndex() != null ? getRequestMessage().getStartingIndex() : 0;
 		if (ignoreCountLimit) {
 			return String.format(sql, column, title, 0, 0);
 		} else {
@@ -190,7 +69,7 @@ public class SearchRequestHandler {
 	 * @return
 	 */
 	private String addSqlSelectByType() {
-		switch (requestType) {
+		switch (getRequestType()) {
 			case TYPE_AUDIO -> {
 				String sql = "SELECT FT.SCORE, A.RATING, A.GENRE, F.FILENAME, F.MODIFIED, F.ID AS FID, FT.SCORE, F.ID AS OID FROM FTL_SEARCH_DATA('%s:%s', %d, %d) FT " +
 					"JOIN AUDIO_METADATA A  ON A.FILEID = FT.KEYS[1] JOIN FILES F ON F.ID = A.FILEID WHERE FT.\"TABLE\" = 'AUDIO_METADATA' AND ";
@@ -238,7 +117,7 @@ public class SearchRequestHandler {
 			case TYPE_VIDEO, TYPE_IMAGE -> {
 				return "select FILENAME, MODIFIED, F.ID as FID, F.ID as oid from FILES as F where ";
 			}
-			default -> throw new RuntimeException("not implemented request type : " + (requestType != null ? requestType : "NULL"));
+			default -> throw new RuntimeException("not implemented request type : " + (getRequestType() != null ? getRequestType() : "NULL"));
 		}
 	}
 
@@ -249,7 +128,7 @@ public class SearchRequestHandler {
 	 * @return
 	 */
 	private String addSqlSelectByType(String subtreeId) {
-		switch (requestType) {
+		switch (getRequestType()) {
 			case TYPE_AUDIO -> {
 				String sql = getTreeStatement(subtreeId) + "SELECT FT.SCORE, A.RATING, A.GENRE, F.FILENAME, F.MODIFIED, F.ID AS FID, F.ID AS OID, FT.SCORE " +
 					"FROM FTL_SEARCH_DATA('%s:%s', %d, %d) FT " +
@@ -307,7 +186,7 @@ public class SearchRequestHandler {
 			case TYPE_VIDEO, TYPE_IMAGE -> {
 				return getTreeStatement(subtreeId) + "select FILENAME, MODIFIED, F.ID as FID, F.ID as oid FROM files " +
 					getTreeWhereStatement("FILES", subtreeId, true);			}
-			default -> throw new RuntimeException("not implemented request type : " + (requestType != null ? requestType : "NULL"));
+			default -> throw new RuntimeException("not implemented request type : " + (getRequestType() != null ? getRequestType() : "NULL"));
 		}
 	}
 
@@ -321,7 +200,7 @@ public class SearchRequestHandler {
 	 * @return
 	 */
 	private String addSqlSelectCountByType() {
-		switch (requestType) {
+		switch (getRequestType()) {
 			case TYPE_AUDIO -> {
 				String sql = "SELECT COUNT(*) FROM FTL_SEARCH_DATA('%s:%s', %d, %d) FT JOIN AUDIO_METADATA A ON A.FILEID = FT.KEYS[1] " +
 					"JOIN FILES F ON F.ID = A.FILEID WHERE FT.\"TABLE\" = 'AUDIO_METADATA' AND ";
@@ -363,7 +242,7 @@ public class SearchRequestHandler {
 			case TYPE_FOLDER -> {
 				return "select count(DISTINCT child.NAME) from STORE_IDS child, STORE_IDS parent where ";
 			}
-			default -> throw new RuntimeException("not implemented request type : " + (requestType != null ? requestType : "NULL"));
+			default -> throw new RuntimeException("not implemented request type : " + (getRequestType() != null ? getRequestType() : "NULL"));
 		}
 	}
 
@@ -375,7 +254,7 @@ public class SearchRequestHandler {
 	 * @return
 	 */
 	private String addSqlSelectCountByType(String subtreeId) {
-		switch (requestType) {
+		switch (getRequestType()) {
 			case TYPE_AUDIO -> {
 				String sql = getTreeStatement(subtreeId) + "SELECT COUNT(*) FROM FTL_SEARCH_DATA('%s:%s', %d, %d) FT " +
 					"JOIN AUDIO_METADATA A ON A.FILEID = FT.KEYS[1] JOIN FILES F ON F.ID = A.FILEID " +
@@ -426,7 +305,7 @@ public class SearchRequestHandler {
 				return getTreeStatement(subtreeId) + "select count(DISTINCT child.NAME) from tree JOIN STORE_IDS child on tree.id = child.id, STORE_IDS parent " +
 					getTreeWhereStatement("FILES", subtreeId, true);
 			}
-			default -> throw new RuntimeException("not implemented request type : " + (requestType != null ? requestType : "NULL"));
+			default -> throw new RuntimeException("not implemented request type : " + (getRequestType() != null ? getRequestType() : "NULL"));
 		}
 	}
 
@@ -449,12 +328,12 @@ public class SearchRequestHandler {
 		}
 		*/
 
-		String title = list.stream().filter(t -> filterAttr.equalsIgnoreCase(t.attr)).findFirst().map(t -> t.val).orElse("");
-		String op = list.stream().filter(t -> filterAttr.equalsIgnoreCase(t.attr)).findFirst().map(t -> t.op).orElse("");
+		String title = list.stream().filter(t -> filterAttr.equalsIgnoreCase(t.attr())).findFirst().map(t -> t.val()).orElse("");
+		String op = list.stream().filter(t -> filterAttr.equalsIgnoreCase(t.attr())).findFirst().map(t -> t.op()).orElse("");
 		// Escape lucene special characters
 		title = LUCENE_PATTERN.matcher(title).replaceAll("\\\\$1");
 		if ("contains".equalsIgnoreCase(op)) {
-			if (umsConfiguration.getLuceneContainsFuzzySearch()) {
+			if (getUmsConfiguration().getLuceneContainsFuzzySearch()) {
 				title = prepareLuceneSearch(title, "~2");
 			} else {
 				if (!(title.startsWith("\"") && title.endsWith("\""))) {
@@ -464,7 +343,7 @@ public class SearchRequestHandler {
 			}
 		}
 		if ("=".equalsIgnoreCase(op)) {
-			if (umsConfiguration.getLuceneEqualFuzzySearch()) {
+			if (getUmsConfiguration().getLuceneEqualFuzzySearch()) {
 				title = prepareLuceneSearch(title, "~2");
 			}
 		}
@@ -473,7 +352,7 @@ public class SearchRequestHandler {
 		if (StringUtils.isBlank(title)) {
 			LOGGER.warn("no search term found for lucene search.");
 			for (SearchToken searchToken : list) {
-				LOGGER.debug("search token : {} {} {}", searchToken.attr, searchToken.op, searchToken.val);
+				LOGGER.debug("search token : {} {} {}", searchToken.attr(), searchToken.op(), searchToken.val());
 			}
 		}
 		return title;
@@ -512,15 +391,15 @@ public class SearchRequestHandler {
 	protected String convertToFilesSql() {
 
 		StringBuilder sb = new StringBuilder();
-		String subtreeId = requestMessage.getContainerId();
+		String subtreeId = getRequestMessage().getContainerId();
 		if ("0".equals(subtreeId) || StringUtils.isAllBlank(subtreeId)) {
 			sb.append(addSqlSelectByType());
 		} else {
 			sb.append(addSqlSelectByType(subtreeId));
 		}
-		addSqlWherePart(requestMessage.getSearchCriteria(), requestType, sb);
-		addOrderBy(requestMessage.getSortCriteria(), requestType, sb);
-		addLimit(requestType, requestMessage, sb);
+		addSqlWherePart(getRequestMessage().getSearchCriteria(), getRequestType(), sb);
+		addOrderBy(getRequestMessage().getSortCriteria(), getRequestType(), sb);
+		addLimit(getRequestType(), getRequestMessage(), sb);
 		LOGGER.debug(sb.toString());
 		return sb.toString();
 	}
@@ -576,44 +455,31 @@ public class SearchRequestHandler {
 				// do nothing, since the FTL search already delivers the correct subset of data based on the startingIndex and requestedCount parameters.
 				}
 			default -> {
-				long limit = requestMessage.getRequestedCount();
+				long limit = getRequestMessage().getRequestedCount();
 				if (limit == 0) {
 					limit = 999; // performance issue: do only deliver top 999 items
 				}
-				sb.append(String.format(" LIMIT %d OFFSET %d ", limit, requestMessage.getStartingIndex()));
+				sb.append(String.format(" LIMIT %d OFFSET %d ", limit, getRequestMessage().getStartingIndex()));
 			}
 		}
 	}
 
-	protected String convertToCountSql(SearchRequest requestMessage) {
+	protected String convertToCountSql() {
 		StringBuilder sb = new StringBuilder();
-		String subtreeId = requestMessage.getContainerId();
-		String upnpSearch = requestMessage.getSearchCriteria();
+		String subtreeId = getRequestMessage().getContainerId();
+		String upnpSearch = getRequestMessage().getSearchCriteria();
 		if ("0".equals(subtreeId) || StringUtils.isAllBlank(subtreeId)) {
 			sb.append(addSqlSelectCountByType());
 		} else {
 			sb.append(addSqlSelectCountByType(subtreeId));
 		}
-		addSqlWherePart(upnpSearch, requestType, sb);
+		addSqlWherePart(upnpSearch, getRequestType(), sb);
 		return sb.toString();
-	}
-
-	private List<SearchToken> getSearchTokens(String searchCriteria) {
-		List<SearchToken> result = new ArrayList<>();
-		Matcher matcher = TOKENIZER_PATTERN.matcher(searchCriteria);
-		while (matcher.find()) {
-			result.add(new SearchToken(matcher.group("property"), matcher.group("op"), makeClean(matcher.group("val"))));
-		}
-		return result;
-	}
-
-	private String makeClean(String val) {
-		return val.replace("\"\"", "\"");
 	}
 
 	private void addSqlWherePart(String searchCriteria, DbIdMediaType requestType, StringBuilder sb) {
 		int lastIndex = 0;
-		Matcher matcher = TOKENIZER_PATTERN.matcher(searchCriteria);
+		Matcher matcher = SearchRequestTokenizer.TOKENIZER_PATTERN.matcher(searchCriteria);
 		while (matcher.find()) {
 			sb.append(searchCriteria, lastIndex, matcher.start());
 			if ("upnp:class".equalsIgnoreCase(matcher.group("property"))) {
@@ -682,83 +548,6 @@ public class SearchRequestHandler {
 		}
 	}
 
-	private String escapeH2dbSql(String val) {
-		val = val.replaceAll("'", "''");
-
-		// Unicode #2018 is send by iOS (since iOS11) if "Smart Punctuation" is
-		// active.
-		val = val.replaceAll("‘", "''");
-		return val;
-	}
-
-	private String getField(String prop, DbIdMediaType requestType) {
-		String property = prop.toLowerCase();
-		if ("dc:title".equalsIgnoreCase(property)) {
-			// handle title by return type.
-			return getTitlePropertyMapping(requestType);
-		} else if (property.startsWith("upnp:artist")) {
-			// check for @role=composer, @role=conductor or @role=albumartist
-			if (property.contains("albumartist")) {
-				return " A.ALBUMARTIST ";
-			} else if (property.contains("composer")) {
-				return " A." + MediaTableAudioMetadata.COL_COMPOSER + " ";
-			} else if (property.contains("conductor")) {
-				return " A." + MediaTableAudioMetadata.COL_CONDUCTOR + " ";
-			}
-			// no role, just the artist
-			return " A.ARTIST ";
-		} else if ("upnp:genre".equals(property)) {
-			return " A.GENRE ";
-		} else if ("dc:creator".equals(property)) {
-			return " A.ALBUMARTIST ";
-		} else if ("upnp:album".equals(property)) {
-			return " A.ALBUM ";
-		} else if ("upnp:rating".equals(property)) {
-			return " rating ";
-		} else if ("ums:likedalbum".equals(property)) {
-			// Makes less sense in a  score based search. Maybe we can use this property in a future implementation to mark albums as
-			// liked in the database and then use this information to boost the score of liked albums in the search results.
-			return " ";
-		} else if ("ums:score".equals(property)) {
-			return " score ";
-		}
-
-		throw new RuntimeException("unknown or unimplemented property: >" + property + "<");
-	}
-
-	private String getTitlePropertyMapping(DbIdMediaType requestType) {
-		switch (requestType) {
-			case TYPE_AUDIO -> {
-				return " A.SONGNAME ";
-			}
-			case TYPE_ALBUM -> {
-				return " A.ALBUM ";
-			}
-			case TYPE_PERSON -> {
-				return " A.ARTIST ";
-			}
-			case TYPE_PERSON_COMPOSER -> {
-				return " A.COMPOSER ";
-			}
-			case TYPE_PERSON_ALBUMARTIST -> {
-				return " A.ALBUMARTIST ";
-			}
-			case TYPE_PERSON_CONDUCTOR -> {
-				return " A.CONDUCTOR ";
-			}
-			case TYPE_PLAYLIST, TYPE_VIDEO, TYPE_IMAGE -> {
-				return " F.FILENAME ";
-			}
-			case TYPE_FOLDER -> {
-				return " child.name ";
-			}
-			default -> {
-				// nothing to do
-			}
-		}
-		throw new RuntimeException("Unknown type : " + requestType);
-	}
-
 	private void acquireDatabaseType(StringBuilder sb, String op, String val, DbIdMediaType requestType) {
 		switch (requestType) {
 			case TYPE_ALBUM, TYPE_PERSON, TYPE_PERSON_COMPOSER, TYPE_PERSON_CONDUCTOR, TYPE_PERSON_ALBUMARTIST, TYPE_FOLDER -> {
@@ -778,229 +567,6 @@ public class SearchRequestHandler {
 		throw new RuntimeException("Unknown type : " + requestType);
 	}
 
-	/**
-	 * unpn:class filetype mapping
-	 *
-	 * @param val
-	 * @return
-	 */
-	private int getFileType(DbIdMediaType mediaFolderType) {
-		// album and persons titles are stored within the RealFile and have
-		// therefore no unique id.
-		switch (mediaFolderType) {
-			case TYPE_AUDIO, TYPE_ALBUM, TYPE_PERSON, TYPE_PERSON_COMPOSER, TYPE_PERSON_CONDUCTOR, TYPE_PERSON_ALBUMARTIST -> {
-				return Format.AUDIO;
-			}
-			case TYPE_VIDEO -> {
-				return Format.VIDEO;
-			}
-			case TYPE_IMAGE -> {
-				return Format.IMAGE;
-			}
-			case TYPE_PLAYLIST -> {
-				return Format.PLAYLIST;
-			}
-			case TYPE_FOLDER -> {
-				// do nothing, where not in the FILES table, but in STORE_IDS
-			}
-			default -> {
-				// nothing to do
-			}
-		}
-		throw new RuntimeException("unknown or unimplemented mediafolder type : >" + mediaFolderType + "<");
-	}
-
-	/**
-	 * Delivers the total count of elements matching the search criteria. This is needed to deliver the correct totalMatches value in the SearchResult.
-	 *
-	 * @param searchRequest
-	 * @return
-	 */
-	public int getSearchCountElements(SearchRequest searchRequest) {
-		String sql = convertToCountSql(searchRequest);
-		return getLibraryResourceCountFromSQL(sql);
-	}
-
-	protected int getLibraryResourceCountFromSQL(String query) {
-		if (LOGGER.isTraceEnabled()) {
-			LOGGER.trace(String.format("SQL count : %s", query));
-		}
-		Connection connection = null;
-		try {
-			connection = MediaDatabase.getConnectionIfAvailable();
-			if (connection != null) {
-				try (Statement statement = connection.createStatement(); ResultSet resultSet = statement.executeQuery(query)) {
-					if (resultSet.next()) {
-						return resultSet.getInt(1);
-					}
-				} catch (SQLException e) {
-					LOGGER.trace("getLibraryResourceCountFromSQL", e);
-				}
-			}
-		} catch (Exception e) {
-			LOGGER.warn("getLibraryResourceCountFromSQL", e);
-		} finally {
-			MediaDatabase.close(connection);
-		}
-		return 0;
-	}
-
-	/**
-	 *  Makes a logical DB search and converts result set to items or container.
-	 *
-	 * @param query
-	 * @return
-	 *
-	 * List of discovered CDS items and containers from the database.
-	 */
-	public List<StoreResource> getLibraryResourceFromSQL(Renderer renderer) {
-		ArrayList<StoreResource> result = new ArrayList<>();
-
-		// SQL statements having 'FILENAME' as a result identifier.
-		String query = convertToFilesSql();
-
-		if (LOGGER.isTraceEnabled()) {
-			LOGGER.trace(String.format("SQL %s : %s", requestType.dbidPrefix, query));
-		}
-		Connection connection = null;
-		try {
-			connection = MediaDatabase.getConnectionIfAvailable();
-			if (connection != null) {
-				try (Statement statement = connection.createStatement()) {
-					try (ResultSet resultSet = statement.executeQuery(query)) {
-						Set<String> foundAlbums = new HashSet<>();
-						while (resultSet.next()) {
-							String filenameField = extractDisplayName(resultSet);
-							switch (requestType) {
-								case TYPE_ALBUM -> {
-									String mbid = resultSet.getString("MBID_RECORD");
-									Long discogs = resultSet.getObject("DISCOGS_RELEASE_ID", Long.class);
-									if (StringUtils.isBlank(mbid) && discogs == null) {
-										// Regular albums can be discovered in the media library
-										StoreResource sr = DbIdResourceLocator.getAlbumFromMediaLibrary(renderer, filenameField);
-										if (sr != null) {
-											result.add(sr);
-										}
-									} else {
-										String identToMatch = null;
-										if (StringUtils.isNotBlank(mbid)) {
-											identToMatch = mbid;
-										} else if (discogs != null) {
-											identToMatch = discogs.toString();
-										}
-										if (!(foundAlbums.contains(identToMatch))) {
-											AlbumMetadata album = new AlbumMetadata(mbid, discogs, resultSet.getString("album"), resultSet.getString("artist"),
-												Integer.toString(resultSet.getInt("media_year")), resultSet.getString("genre"));
-											MusicAlbumFolder folder = DbIdResourceLocator.getLibraryResourceMusicBrainzFolder(
-												renderer, album.getTypeIdent());
-											if (folder == null) {
-												folder = DbIdLibrary.addLibraryResourceMusicAlbum(renderer, album);
-											}
-											result.add(folder);
-											foundAlbums.add(identToMatch);
-										}
-									}
-								}
-								case TYPE_PERSON, TYPE_PERSON_COMPOSER, TYPE_PERSON_CONDUCTOR, TYPE_PERSON_ALBUMARTIST -> {
-									DbIdTypeAndIdent ti = new DbIdTypeAndIdent(requestType, filenameField);
-									MusicBrainzPersonFolder personFolder = DbIdResourceLocator.getLibraryResourcePersonFolder(renderer, ti);
-									if (personFolder == null) {
-										personFolder = DbIdLibrary.addLibraryResourcePerson(renderer, ti);
-									}
-									result.add(personFolder);
-								}
-								case TYPE_PLAYLIST -> {
-									String realFileName = resultSet.getString("FILENAME");
-									if (realFileName != null) {
-										StoreResource res = DbIdResourceLocator.getLibraryResourcePlaylist(renderer, realFileName);
-										if (res != null) {
-											result.add(res);
-										}
-									}
-								}
-								case TYPE_FOLDER -> {
-									if (filenameField != null) {
-										StoreResource res = DbIdResourceLocator.getLibraryResourceFolder(renderer, filenameField);
-										if (res != null) {
-											result.add(res);
-										}
-									}
-								}
-								default -> {
-									String realFileName = resultSet.getString("FILENAME");
-									if (realFileName != null) {
-										StoreResource res = DbIdResourceLocator.getLibraryResourceRealFile(renderer, realFileName);
-										if (res != null) {
-											res.resolve();
-											result.add(res);
-										}
-									}
-								}
-							}
-						}
-					}
-				}
-			}
-		} catch (SQLException e) {
-			LOGGER.warn("getLibraryResourceFromSQL", e);
-		} finally {
-			MediaDatabase.close(connection);
-		}
-		return result;
-	}
-
-	private String extractDisplayName(ResultSet resultSet) throws SQLException {
-		switch (requestType) {
-			case TYPE_VIDEO, TYPE_PLAYLIST, TYPE_IMAGE, TYPE_AUDIO -> {
-				return FilenameUtils.getBaseName(resultSet.getString("FILENAME"));
-			}
-			case TYPE_FOLDER -> {
-				return resultSet.getString("name");
-			}
-			default -> {
-				// artificial field 'filename' of a person or similar type is
-				// already the final display name.
-				return resultSet.getString("FILENAME");
-			}
-		}
-	}
-
-	/**
-	 * Wraps the payload around soap Envelope / Body tags.
-	 *
-	 * @param payload Soap body as a XML String
-	 * @return Soap message as a XML string
-	 */
-	private StringBuilder createResponse(String payload) {
-		StringBuilder response = new StringBuilder();
-		response.append(HTTPXMLHelper.XML_HEADER).append(CRLF);
-		response.append(HTTPXMLHelper.SOAP_ENCODING_HEADER).append(CRLF);
-		response.append(payload).append(CRLF);
-		response.append(HTTPXMLHelper.SOAP_ENCODING_FOOTER).append(CRLF);
-		return response;
-	}
-
-	private StringBuilder buildEnvelope(int foundNumberReturned, int totalMatches, long updateID, StringBuilder dlnaItems) {
-		StringBuilder response = new StringBuilder();
-		response.append(HTTPXMLHelper.SEARCHRESPONSE_HEADER);
-		response.append(CRLF);
-		response.append(HTTPXMLHelper.RESULT_HEADER);
-		response.append(HTTPXMLHelper.DIDL_HEADER);
-		response.append(dlnaItems.toString());
-		response.append(HTTPXMLHelper.DIDL_FOOTER);
-		response.append(HTTPXMLHelper.RESULT_FOOTER);
-		response.append(CRLF);
-		response.append("<NumberReturned>").append(foundNumberReturned).append("</NumberReturned>");
-		response.append(CRLF);
-		response.append("<TotalMatches>").append(totalMatches).append("</TotalMatches>");
-		response.append(CRLF);
-		response.append("<UpdateID>");
-		response.append(updateID);
-		response.append("</UpdateID>");
-		response.append(CRLF);
-		response.append(HTTPXMLHelper.SEARCHRESPONSE_FOOTER);
-		return response;
-	}
 
 	private String getTreeStatement(String subtreeId) {
 		String tree = String.format("WITH RECURSIVE tree(id, name) AS (\n" +
