+++ src/main/java/net/pms/network/mediaserver/jupnp/support/contentdirectory/UmsContentDirectoryService.java
@@ -28,30 +28,6 @@ import java.util.TimerTask;
 import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.transform.TransformerException;
 import javax.xml.xpath.XPathExpressionException;
-import net.pms.dlna.DidlHelper;
-import net.pms.network.mediaserver.handlers.SearchRequestHandler;
-import net.pms.network.mediaserver.handlers.message.SearchRequest;
-import net.pms.network.mediaserver.jupnp.model.meta.UmsRemoteClientInfo;
-import net.pms.network.mediaserver.jupnp.support.contentdirectory.result.Parser;
-import net.pms.network.mediaserver.jupnp.support.contentdirectory.result.Result;
-import net.pms.network.mediaserver.jupnp.support.contentdirectory.result.StoreResourceHelper;
-import net.pms.network.mediaserver.jupnp.support.contentdirectory.result.namespace.didl_lite.container.Container;
-import net.pms.network.mediaserver.jupnp.support.contentdirectory.result.namespace.didl_lite.item.Item;
-import net.pms.network.mediaserver.jupnp.support.contentdirectory.updateobject.IUpdateObjectHandler;
-import net.pms.network.mediaserver.jupnp.support.contentdirectory.updateobject.UpdateObjectFactory;
-import net.pms.renderers.Renderer;
-import net.pms.store.MediaScanner;
-import net.pms.store.MediaStatusStore;
-import net.pms.store.MediaStoreIds;
-import net.pms.store.PlaylistManager;
-import net.pms.store.StoreContainer;
-import net.pms.store.StoreItem;
-import net.pms.store.StoreResource;
-import net.pms.store.container.MediaLibrary;
-import net.pms.store.container.PlaylistFolder;
-import net.pms.store.utils.StoreResourceSorter;
-import net.pms.util.StringUtil;
-import net.pms.util.UMSUtils;
 import org.jupnp.binding.annotations.UpnpAction;
 import org.jupnp.binding.annotations.UpnpInputArgument;
 import org.jupnp.binding.annotations.UpnpOutputArgument;
@@ -74,6 +50,32 @@ import org.jupnp.support.model.SortCriterion;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.xml.sax.SAXException;
+import net.pms.dlna.DidlHelper;
+import net.pms.network.mediaserver.handlers.BaseSearchRequestHandler;
+import net.pms.network.mediaserver.handlers.LucenseSearchRequestHandler;
+import net.pms.network.mediaserver.handlers.SearchRequestTokenizer;
+import net.pms.network.mediaserver.handlers.message.SearchRequest;
+import net.pms.network.mediaserver.jupnp.model.meta.UmsRemoteClientInfo;
+import net.pms.network.mediaserver.jupnp.support.contentdirectory.result.Parser;
+import net.pms.network.mediaserver.jupnp.support.contentdirectory.result.Result;
+import net.pms.network.mediaserver.jupnp.support.contentdirectory.result.StoreResourceHelper;
+import net.pms.network.mediaserver.jupnp.support.contentdirectory.result.namespace.didl_lite.container.Container;
+import net.pms.network.mediaserver.jupnp.support.contentdirectory.result.namespace.didl_lite.item.Item;
+import net.pms.network.mediaserver.jupnp.support.contentdirectory.updateobject.IUpdateObjectHandler;
+import net.pms.network.mediaserver.jupnp.support.contentdirectory.updateobject.UpdateObjectFactory;
+import net.pms.renderers.Renderer;
+import net.pms.store.MediaScanner;
+import net.pms.store.MediaStatusStore;
+import net.pms.store.MediaStoreIds;
+import net.pms.store.PlaylistManager;
+import net.pms.store.StoreContainer;
+import net.pms.store.StoreItem;
+import net.pms.store.StoreResource;
+import net.pms.store.container.MediaLibrary;
+import net.pms.store.container.PlaylistFolder;
+import net.pms.store.utils.StoreResourceSorter;
+import net.pms.util.StringUtil;
+import net.pms.util.UMSUtils;
 
 @UpnpService(
 		serviceId =
@@ -872,7 +874,13 @@ public class UmsContentDirectoryService {
 		}
 
 		try {
-			SearchRequestHandler searchRequestHandler = new SearchRequestHandler(searchRequest);
+			BaseSearchRequestHandler searchRequestHandler = null;
+			SearchRequestTokenizer tokenizer = new SearchRequestTokenizer(searchRequest);
+			if (tokenizer.hasDcTitleSearch()) {
+				searchRequestHandler = new LucenseSearchRequestHandler(searchRequest);
+			} else {
+
+			}
 
 			int totalMatches = searchRequestHandler.getSearchCountElements(searchRequest);
 			LOGGER.debug("{}", searchRequest.getSearchCriteria());
